broke main_methods.py into smaller, more specific modules
